[move/#63] usecase auth 파일 이동
